Moved heatmap stuff into separate file
