+++ atmosphere.py
@@ -7,10 +7,11 @@ from mcm_utils import *
 import matplotlib as mpl
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib.pyplot as plt
+from heatmap import heatmap
 
 class atmosphere:
     def __init__(self):
-        self.number_rays = 1000
+        self.number_rays = 10000
         self.ray_start = [200,0,0]
         self.ray_direction = [1,1,0]
    
@@ -95,28 +96,6 @@ class atmosphere:
         [ax.plot(*vals,color='b',alpha=4/len(data)) for vals in data]
 
         plt.show()
-
-class heatmap:
-    def __init__(self):
-        self.intensity=np.zeros((30,30))
-        longitudes=np.linspace(0,360,num=30,endpoint=False)
-        extended_longitudes=np.tile(longitudes,(30,1))
-        sin_latitudes=np.linspace(1,-1,num=30,endpoint=False)
-        extended_sin_latitudes=np.tile(sin_latitudes,(30,1)).T
-        self.upper_right=np.swapaxes(np.array([extended_longitudes,extended_sin_latitudes]),0,2)
-        self.rights=extended_longitudes
-        self.uppers=extended_sin_latitudes
-
-    def update_regions(self,intersection_points):
-        self.reset()
-        latlongs=geographic_coordinates(intersection_points).T
-        xreg=(latlongs[1]//12).astype(int)
-        yreg=((1-np.sin(deg2rad(latlongs[0])))//(2/30)).astype(int)
-        for k in range(xreg.shape[0]):
-                self.intensity[xreg[k],yreg[k]]+=1            
-        
-    def reset(self):
-        self.intensity=np.zeros((30,30))    
 
 if __name__=="__main__":
     world = atmosphere()
